merge to MASTER => v0.4.0

+++ app/webFrontend/src/app/unit/unit-form/task-unit-edit/task-unit-edit.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input, OnInit, ViewChild} from '@angular/core';
 import {TaskService, UnitService} from '../../../shared/services/data.service';
 import {Task} from '../../../models/Task';
-import {MdSnackBar} from '@angular/material';
+import {MatSnackBar} from '@angular/material';
 import {ITaskUnit} from '../../../../../../../shared/models/units/ITaskUnit';
 import {TaskUnit} from '../../../models/TaskUnit';
 import {ITask} from '../../../../../../../shared/models/task/ITask';
@@ -30,76 +30,108 @@ export class TaskUnitEditComponent implements OnInit {
   @Input()
   onCancel: () => void;
 
+  add = false;
+
   @ViewChild(UnitGeneralInfoFormComponent)
   private generalInfo: UnitGeneralInfoFormComponent;
 
-  tasks: any[];
-
   constructor(private taskService: TaskService,
               private unitService: UnitService,
-              private snackBar: MdSnackBar) {
+              private snackBar: MatSnackBar) {
   }
 
   ngOnInit() {
-    console.log(this.model);
-
     if (!this.model) {
       this.model = new TaskUnit(this.course._id);
+      this.add = true;
+    } else {
+      this.reloadTaskUnit();
     }
-
-    this.loadTasksFromServer();
   }
 
-  loadTasksFromServer() {
-    this.taskService.getTasksForCourse(this.course._id).then(tasks => {
-      this.tasks = tasks;
-
-      for (const task of this.tasks) {
-        this.answerPreparationAfterLoadingFromServer(task); // WORKAROUND get rid of the _id for the answers
-      }
-    });
+  async reloadTaskUnit() {
+    // Reload the task unit from the database to make sure that the tasks (and answers)
+    // are populated properly (e.g. necessary after a Cancel)
+    this.model = <ITaskUnit><any>await this.unitService.readTaskUnit(this.model._id);
   }
 
-  addUnit() {
-    this.unitService.addTaskUnit({
+  saveUnit() {
+    this.model = {
+      ...this.model,
       name: this.generalInfo.form.value.name,
       description: this.generalInfo.form.value.description,
-      ...this.model
-    }, this.lectureId)
-    .then(
-      (task) => {
-        this.snackBar.open('Task created', '', {duration: 3000});
-        this.onDone();
-      },
-      (error) => {
-        console.log(error);
-      });
+      deadline: this.generalInfo.form.value.deadline,
+    };
+
+    if (this.isTaskUnitValid()) {
+      let taskPromise = null;
+      if (this.add) {
+        taskPromise = this.unitService.addTaskUnit(this.model, this.lectureId);
+      } else {
+        taskPromise = this.unitService.updateTaskUnit(this.model);
+      }
+      taskPromise.then(
+        (task) => {
+          const message = `Task ${this.add ? 'created' : 'updated'}`;
+          this.snackBar.open(message, '', {duration: 3000});
+          this.onDone();
+        },
+        (error) => {
+          const message = `Couldn\'t ${this.add ? 'create' : 'update'} task`;
+          this.snackBar.open(message, '', {duration: 3000});
+        });
+    }
   };
+
+  isTaskUnitValid() {
+    const taskUnit = this.model;
+
+    if (taskUnit.name === null || taskUnit.name.trim() === '') {
+      const message = 'Task not valid: Name is required';
+      this.snackBar.open(message, '', {duration: 3000});
+      return false
+    } else if (taskUnit.tasks.length === 0) {
+      const message = 'Task not valid: At least one question is required';
+      this.snackBar.open(message, '', {duration: 3000});
+      return false
+    } else {
+      // Check if all tasks i.e. questions are valid
+      for (const task of taskUnit.tasks) {
+        if (task.name === undefined || task.name.trim() === '') {
+          const message = 'Task not valid: Every question requires some text';
+          this.snackBar.open(message, '', {duration: 3000});
+          return false
+        } else if (task.answers.length < 2) {
+          const message = 'Task not valid: Every question requires at least two answers';
+          this.snackBar.open(message, '', {duration: 3000});
+          return false
+        } else {
+          // Check if answers are valid
+          let noAnswersChecked = true;
+
+          for (const answer of task.answers) {
+            if (noAnswersChecked) {
+              noAnswersChecked = !answer.value
+            }
+            if (answer.text === undefined || answer.text.trim() === '') {
+              const message = 'Task not valid: Every answer requires some text';
+              this.snackBar.open(message, '', {duration: 3000});
+              return false
+            }
+          }
+          if (noAnswersChecked) {
+            const message = 'Task not valid: Every question requires at least one checked answer';
+            this.snackBar.open(message, '', {duration: 3000});
+            return false
+          }
+        }
+      }
+    }
+    return true
+  }
 
   addTask() {
     this.model.tasks.push(new Task());
-    // this.createTask(newTask);
-  }
-
-  //  log(val) { console.log(JSON.stringify(val)); }
-
-  createTask(task: any) {
-    // this.log(this.task);
-    this.taskService.createItem(task).then(
-      (val) => {
-        task = val; // get _id
-        this.tasks.splice(0, 0, task); // add item to start
-
-        //     this.log(val);
-      }, (error) => {
-        console.log(error);
-      });
-  }
-
-  answerPreparationAfterLoadingFromServer(task: any) {
-    for (const answer of task.answers) {
-      delete answer._id;
-    }
   }
 
   addAnswerAtEnd(task: ITask) {
@@ -110,22 +142,15 @@ export class TaskUnitEditComponent implements OnInit {
     task.answers.pop();
   }
 
-  updateTask(task: any) {
-    // this.log(this.task);
-    this.taskService.updateItem(task).then(
-      (val) => {
-        this.snackBar.open('Task saved', 'Update', {duration: 2000});
-
-      }, (error) => {
-        console.log(error);
-      });
-  }
-
   removeTask(task: any) {
-    this.taskService.deleteItem(task).then(tasks => {
-      this.tasks = (this.tasks.filter(obj => task._id !== obj._id));
-      this.snackBar.open('Task deleted', 'Delete', {duration: 2000});
+    if (!task._id) {
+      this.model.tasks = this.model.tasks.filter(obj => task !== obj);
+      return;
+    }
 
+    this.taskService.deleteItem(task).then(tasks => {
+      this.model.tasks = this.model.tasks.filter(obj => task._id !== obj._id);
+      this.snackBar.open('Task deleted', 'Delete', {duration: 2000});
     });
   }
 }
